runtime: clean up realtime impl namespace
(cherry picked from commit 032ac81a4cddc36f7d184500a5bb2e7d8419014b)
