+++ gnuradio-runtime/include/gnuradio/CMakeLists.txt
@@ -44,7 +44,6 @@ install(FILES
   pycallback_object.h
   random.h
   realtime.h
-  realtime_impl.h
   runtime_types.h
   tags.h
   tagged_stream_block.h
