+++ gnuradio-runtime/include/gnuradio/realtime.h
@@ -12,9 +12,15 @@
 #define INCLUDED_GR_REALTIME_H
 
 #include <gnuradio/api.h>
-#include <gnuradio/realtime_impl.h>
 
 namespace gr {
+
+typedef enum { RT_OK = 0, RT_NOT_IMPLEMENTED, RT_NO_PRIVS, RT_OTHER_ERROR } rt_status_t;
+
+enum rt_sched_policy {
+    RT_SCHED_RR = 0,   // round robin
+    RT_SCHED_FIFO = 1, // first in first out
+};
 
 /*!
  * \brief If possible, enable high-priority "real time" scheduling.
