+++ gnuradio-runtime/lib/realtime_impl.h
@@ -12,43 +12,36 @@
 #define INCLUDED_GNURADIO_REALTIME_H
 
 #include <gnuradio/api.h>
+#include <gnuradio/realtime.h>
 #include <stdexcept>
 
 /*!
  * \brief System independent way to ask for realtime scheduling
  */
 namespace gr {
-
-typedef enum { RT_OK = 0, RT_NOT_IMPLEMENTED, RT_NO_PRIVS, RT_OTHER_ERROR } rt_status_t;
-
-enum rt_sched_policy {
-    RT_SCHED_RR = 0,   // round robin
-    RT_SCHED_FIFO = 1, // first in first out
-};
-
-namespace impl {
+namespace realtime {
 /*
  * Define the range for our virtual priorities (don't change
  * these)
  *
  * Processes (or threads) with numerically higher priority values
  * are scheduled before processes with numerically lower priority
- * values. Thus, the value returned by rt_priority_max() will be
- * greater than the value returned by rt_priority_min().
+ * values. Thus, the value returned by s_rt_priority_max will be
+ * greater than the value returned by s_rt_priority_min.
  */
-static inline int rt_priority_min() { return 0; }
-static inline int rt_priority_max() { return 15; }
-static inline int rt_priority_default() { return 1; }
+static const int s_rt_priority_min = 0;
+static const int s_rt_priority_max = 15;
+static const int s_rt_priority_default = 1;
 
 struct GR_RUNTIME_API rt_sched_param {
     int priority;
     rt_sched_policy policy;
 
-    rt_sched_param() : priority(rt_priority_default()), policy(RT_SCHED_RR) {}
+    rt_sched_param() : priority(s_rt_priority_default), policy(RT_SCHED_RR) {}
 
     rt_sched_param(int priority_, rt_sched_policy policy_ = RT_SCHED_RR)
     {
-        if (priority_ < rt_priority_min() || priority_ > rt_priority_max())
+        if (priority_ < s_rt_priority_min || priority_ > s_rt_priority_max)
             throw std::invalid_argument("rt_sched_param: priority out of range");
 
         priority = priority_;
@@ -68,7 +61,7 @@ struct GR_RUNTIME_API rt_sched_param {
 
 GR_RUNTIME_API rt_status_t enable_realtime_scheduling(rt_sched_param = rt_sched_param());
 
-} /* namespace impl */
+} /* namespace realtime */
 } /* namespace gr */
 
 #endif /* INCLUDED_GNURADIO_REALTIME_H */
